Give send-to-repl file and line number metadata.
The compiler uses these for tagging stack traces and vars.  The latter
will be ultimately useful for implementing a jump to definition
feature.

Unfortunately LineNumberingPushbackReader does not expose
setLineNumber() so we have to proxy getLineNumber() and create a var
which holds an offset that is added to the current line number.

We use set! rather than wrapping the code to be evaled in a binding
so that evaled forms remain at the top-level.  This is necessary for
some forms like 'use to behave correctly.

This patch does not include usages for send-to-repl, it just adds the
arguments.  Usages will be added in a following patch.